requirements.txt
fastapi
uvicorn
motor
pydantic
email-validator
passlib[bcrypt]
pyjwt
python-jose
python-multipart
slowapi
fastapi-csrf-protect